src/pyproject.toml
[tool.poetry]
name = "bcraft"
version = "0.1.0"
description = ""
authors = ["krestovsky13"]
readme = "README.md"
packages = [{include = "bcraft", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
sqlalchemy = "^2.0.7"
alembic = "^1.10.2"
psycopg2 = "^2.9.5"
python-dotenv = "^1.0.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
